requirements.txt
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
decorator==5.2.1
fastjsonschema==2.21.2
Flask==3.1.3
fonttools==4.61.1
fredapi==0.5.2
frozendict==2.4.7
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
multitasking==0.0.12
narwhals==2.17.0
nbformat==5.10.4
numpy==2.4.2
numpy-financial==1.0.0
packaging==26.0
pandas==3.0.1
pandas-datareader==0.10.0
patsy==1.0.2
peewee==4.0.1
pillow==12.1.1
platformdirs==4.9.2
plotly==6.6.0
protobuf==7.34.0
pyarrow==23.0.1
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.2.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
statsmodels==0.14.6
threadpoolctl==3.6.0
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.6
yfinance==1.2.0
